--- frontend/src/metabase/common/components/FilterPicker/FilterFooter/index.ts
@@ -1,1 +1,0 @@
-export * from "./FilterFooter";
