--- frontend/src/metabase/common/components/FilterPicker/FilterHeader/index.ts
@@ -1,1 +1,0 @@
-export * from "./FilterHeader";
