+++ frontend/src/metabase/common/components/FilterPicker/FilterPickerFooter/index.ts
@@ -1,0 +1,1 @@
+export * from "./FilterPickerFooter";
