+++ frontend/src/metabase/common/components/FilterPicker/FilterPickerHeader/FilterPickerHeader.styled.tsx
