+++ frontend/src/metabase/common/components/FilterPicker/FilterPickerHeader/index.ts
@@ -1,0 +1,1 @@
+export * from "./FilterPickerHeader";
